Very basic downloading using ffmpeg and still no packet fixes
notes about how to fix the packets are in the readme and I will work on
raw packet downloading or using a proxy to intercept the packets but I
don't think this is going to work as well as just using the raw (edited
packets) for downloading and playing. Basically the whole thing needs a
rewrite after the link is fetched.